requirements.txt

uvicorn == 0.30.0
# py-hot-reload == 1.0.6
# meson-python == 0.16.0
# numpy == 2.0.0
opencv-python == 4.10.0.84
flask == 3.0.3
flask_socketio == 5.3.6
# fastapi == 0.111.0
# python-socketio == 5.11.3
# jinja2 == 3.1.4
gunicorn == 22.0.0
Flask-Cors == 4.0.1
mediapipe == 0.10.14
cvzone == 1.6.1
pyserial == 3.5
